Create project cli(AIGEAR-16)

+++ src/aigear/project.py
@@ -5,7 +5,7 @@ import shutil
 class Project:
     def __init__(self, name: str = "template_project", version: str = "0.0.1"):
         """
-        Set projects and use versions
+        Used to manage project related functions and information
 
         name: project name
         version: Project version to manage infrastructure
@@ -20,4 +20,6 @@ class Project:
         template_path = os.path.abspath(os.path.dirname(__file__))
         template_folder = os.path.join(template_path, "template")
         prodict_folder = os.path.join(os.getcwd(), self.name)
-        shutil.copytree(template_folder, prodict_folder)
+
+        if not os.path.exists(prodict_folder):
+            shutil.copytree(template_folder, prodict_folder)
